# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stackr API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/domain-resources": {
      "get": {
        "summary": "List domain resources (scope: stacks:read)",
        "operationId": "listDomainResources",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1DomainResourceOut"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "post": {
        "summary": "Create domain resource (scope: domains:write)",
        "operationId": "createDomainResource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1DomainResourceIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1DomainResourceOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/api/v1/domain-resources/{id}": {
      "delete": {
        "summary": "Delete domain resource (scope: domains:write)",
        "operationId": "deleteDomainResource",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "patch": {
        "summary": "Set a domain resource's ACME account (scope: domains:write)",
        "operationId": "setADomainResource'sAcmeAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1DomainResourcePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1DomainResourceOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/api/v1/storage": {
      "get": {
        "summary": "List storage (scope: storage:read)",
        "operationId": "listStorage",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1StorageOut"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "post": {
        "summary": "Create storage (probes on create) (scope: storage:write)",
        "operationId": "createStorage(probesOnCreate)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1StorageIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1StorageOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/api/v1/volumes/{id}": {
      "delete": {
        "summary": "Remove volume (scope: apps:write)",
        "operationId": "removeVolume",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "V1DomainResourceIn": {
        "required": [
          "host"
        ],
        "type": "object",
        "properties": {
          "acme_email": {
            "type": "string",
            "description": "Let's Encrypt account for certificates under this host; empty uses the instance's"
          },
          "host": {
            "type": "string",
            "description": "bare lowercase hostname, e.g. example.com"
          },
          "include_env_on_default": {
            "type": "boolean",
            "description": "keep the default env's slug in generated names"
          },
          "level": {
            "type": "string",
            "description": "instance or node (default, admins only) | org | stack"
          },
          "owner": {
            "type": "string",
            "description": "org or stack id the resource belongs to; ignored for instance level"
          }
        }
      },
      "V1DomainResourceOut": {
        "type": "object",
        "properties": {
          "acme_email": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "include_env_on_default": {
            "type": "boolean"
          },
          "level": {
            "type": "string"
          },
          "owner_id": {
            "type": "string"
          }
        }
      },
      "V1DomainResourcePatch": {
        "type": "object",
        "properties": {
          "acme_email": {
            "type": "string",
            "description": "empty uses the instance's account",
            "nullable": true
          }
        }
      },
...
      "V1ErrBody": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/V1ErrDetail"
          }
        }
      },
      "V1ErrDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "V1StorageIn": {
        "required": [
          "name",
          "backend"
        ],
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "host, for nfs/smb"
          },
          "backend": {
            "type": "string",
            "description": "nfs | smb | local"
          },
          "export": {
            "type": "string",
            "description": "/export (nfs) | share (smb) | absolute host path (local)"
          },
          "name": {
            "type": "string"
          },
          "opts": {
            "type": "string",
            "description": "extra mount opts"
          },
          "org": {
            "type": "string",
            "description": "org slug: makes an org network share (nfs/smb) its tiles mount as ${{ org.storage.NAME }}"
          },
          "password": {
            "type": "string",
            "description": "smb; encrypted at rest (docker volume metadata keeps a plaintext copy)"
          },
          "username": {
            "type": "string",
            "description": "smb"
          }
        }
      },
      "V1StorageOut": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "backend": {
            "type": "string"
          },
          "export": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "org": {
            "type": "string",
            "description": "org slug, set on an org share"
          },
          "paths": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/V1StoragePathOut"
            },
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "status_msg": {
            "type": "string"
          }
        }
      },
...
      "V1StoragePathOut": {
        "type": "object",
        "properties": {
          "forced_ro": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "subpath": {
            "type": "string"
          },
          "volume": {
            "type": "string",
            "description": "backing docker volume name"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header",
        "description": "API key"
      }
    }
  }
}